backend/requirements.txt
# Kite backend dependencies
# Install with: pip install -r requirements.txt

# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Environment variables
python-dotenv==1.0.1

# Database — Supabase client (for auth/REST) + psycopg2 for direct pgvector access
supabase==2.10.0
psycopg2-binary==2.9.10
pgvector==0.3.5

# Embeddings — local model, no API cost
sentence-transformers==3.3.1
numpy==1.26.4

# LLM
anthropic==0.40.0

# SEC filings
sec-edgar-downloader==5.1.0
beautifulsoup4==4.12.3
lxml==5.3.0

# Tokenizer — for chunking
tiktoken==0.8.0

# Market data
yfinance==0.2.50

# News
newsapi-python==0.2.7

# Scheduler for daily cron jobs
apscheduler==3.10.4

# HTTP client (used in ingestion modules)
httpx==0.27.2

# Data validation
pydantic==2.9.2
